+++ Packages/com.texelsaur.video/Runtime/Scripts/Component/SRTCaptions.cs
@@ -277,7 +277,6 @@ at commonsensemedia.org.
                 return;
 
             position = Time.time - positionSyncOffset;
-            Debug.Log($"Position {position} index {index}");
 
             _UpdateIndex();
             _UpdateDisplay();